refactor(onboarding): rename Integrations component folder
Use the broader components/Integrations/ name (was IntegrationsDrawer/) so the
module can house more integration-related surfaces beyond the drawer.
