+++ packages/app/src/components/Integrations/GuideView.tsx
